further reduce circular dependencies
