+++ src/main/java/de/tum/in/www1/artemis/service/AuthorizationCheckService.java
@@ -1,5 +1,7 @@
 package de.tum.in.www1.artemis.service;
 
+import static de.tum.in.www1.artemis.security.AuthoritiesConstants.*;
+
 import java.security.Principal;
 import java.time.ZonedDateTime;
 import java.util.Optional;
