+++ src/main/java/de/tum/in/www1/artemis/service/user/UserService.java
@@ -51,9 +51,6 @@ public class UserService {
 
     private final Logger log = LoggerFactory.getLogger(UserService.class);
 
-    @Value("${artemis.user-management.external.admin-group-name:#{null}}")
-    private Optional<String> adminGroupName;
-
     @Value("${artemis.user-management.use-external}")
     private Boolean useExternalUserManagement;
 
@@ -76,9 +73,11 @@ public class UserService {
 
     private final PasswordService passwordService;
 
+    private final AuthorityService authorityService;
+
     private final Optional<LdapUserService> ldapUserService;
 
-    private Optional<VcsUserManagementService> optionalVcsUserManagementService;
+    private final Optional<VcsUserManagementService> optionalVcsUserManagementService;
 
     private ArtemisAuthenticationProvider artemisAuthenticationProvider;
 
@@ -90,25 +89,21 @@ public class UserService {
 
     private final GuidedTourSettingsRepository guidedTourSettingsRepository;
 
-    public UserService(UserRepository userRepository, AuthorityRepository authorityRepository, CacheManager cacheManager, Optional<LdapUserService> ldapUserService,
-            GuidedTourSettingsRepository guidedTourSettingsRepository, CourseRepository courseRepository, PasswordService passwordService) {
+    public UserService(UserRepository userRepository, AuthorityService authorityService, AuthorityRepository authorityRepository, CacheManager cacheManager,
+            Optional<LdapUserService> ldapUserService, GuidedTourSettingsRepository guidedTourSettingsRepository, CourseRepository courseRepository,
+            PasswordService passwordService, Optional<VcsUserManagementService> optionalVcsUserManagementService) {
         this.userRepository = userRepository;
+        this.authorityService = authorityService;
         this.authorityRepository = authorityRepository;
         this.cacheManager = cacheManager;
         this.ldapUserService = ldapUserService;
         this.guidedTourSettingsRepository = guidedTourSettingsRepository;
         this.courseRepository = courseRepository;
         this.passwordService = passwordService;
-    }
-
-    @Autowired
-    // break the dependency cycle
-    public void setOptionalVcsUserManagementService(Optional<VcsUserManagementService> optionalVcsUserManagementService) {
         this.optionalVcsUserManagementService = optionalVcsUserManagementService;
     }
 
-    @Autowired
-    // break the dependency cycle
+    @Autowired // break the dependency cycle
     public void setArtemisAuthenticationProvider(ArtemisAuthenticationProvider artemisAuthenticationProvider) {
         this.artemisAuthenticationProvider = artemisAuthenticationProvider;
     }
@@ -707,7 +702,7 @@ public class UserService {
         log.debug("Add user " + user.getLogin() + " to group " + group);
         if (!user.getGroups().contains(group)) {
             user.getGroups().add(group);
-            user.setAuthorities(buildAuthorities(user));
+            user.setAuthorities(authorityService.buildAuthorities(user));
             saveUser(user);
         }
     }
@@ -788,54 +783,8 @@ public class UserService {
         log.info("Remove user " + user.getLogin() + " from group " + group);
         if (user.getGroups().contains(group)) {
             user.getGroups().remove(group);
-            user.setAuthorities(buildAuthorities(user));
+            user.setAuthorities(authorityService.buildAuthorities(user));
             saveUser(user);
         }
-    }
-
-    /**
-     * Builds the authorities list from the groups:
-     * <p>
-     * 1) Admin group if the globally defined ADMIN_GROUP_NAME is available and is contained in the users groups, or if the user was an admin before
-     * 2) group contains configured instructor group name -> instructor role
-     * 3) group contains configured tutor group name -> tutor role
-     * 4) the user role is always given
-     *
-     * @param user a user with groups
-     * @return a set of authorities based on the course configuration and the given groups
-     */
-    public Set<Authority> buildAuthorities(User user) {
-        Set<Authority> authorities = new HashSet<>();
-        Set<String> groups = user.getGroups();
-        if (groups == null) {
-            // prevent null pointer exceptions
-            groups = new HashSet<>();
-        }
-
-        // Check if the user is admin in case the admin group is defined
-        if (adminGroupName.isPresent() && groups.contains(adminGroupName.get())) {
-            authorities.add(ADMIN_AUTHORITY);
-        }
-
-        // Users who already have admin access, keep admin access.
-        if (user.getAuthorities() != null && user.getAuthorities().contains(ADMIN_AUTHORITY)) {
-            authorities.add(ADMIN_AUTHORITY);
-        }
-
-        Set<String> instructorGroups = courseRepository.findAllInstructorGroupNames();
-        Set<String> teachingAssistantGroups = courseRepository.findAllTeachingAssistantGroupNames();
-
-        // Check if user is an instructor in any course
-        if (groups.stream().anyMatch(instructorGroups::contains)) {
-            authorities.add(new Authority(INSTRUCTOR));
-        }
-
-        // Check if user is a tutor in any course
-        if (groups.stream().anyMatch(teachingAssistantGroups::contains)) {
-            authorities.add(new Authority(TEACHING_ASSISTANT));
-        }
-
-        authorities.add(new Authority(USER));
-        return authorities;
     }
 }
